Add attribute values to an existing element in a given PMML file.

Add attribute values to an existing element in a given PMML file.

add_attributes(xml_model = NULL, xpath = NULL, attributes = NULL,
  namespace = "4_3", ...)

Arguments

xml_model
The PMML model in a XML node format. If the model is a text file, it should be converted to an XML node, for example, using the file_to_xml_node function.

xpath
The XPath to the element to which the attributes are to be added.

attributes
The attributes to be added to the data fields. The user should make sure that the attributes being added are allowed in the PMML schema.

namespace
The namespace of the PMML model. This is frequently also the PMML version of the model.

…
Further arguments passed to or from other methods.

Value

An object of class XMLNode as that defined by the XML package. This represents the top level, or root node, of the XML document and is of type PMML. It can be written to file with saveXML.

Details

Add attributes to an arbitrary XML element. This is an experimental function designed to be more general than the 'add_mining_field_attributes' and 'add_data_field_attriubutes' functions.

The attribute information can be provided as a vector. Multiple attribute names and values can be passes as vector elements to enable inserting multiple attributes. However, this function overwrites any pre-existing attribute values, so it must be used with care. This behavior is by design as this feature is meant to help an user add new defined attribute values at different times. The XPath has to include the namespace as shown in the examples.

Examples

# Make a sample model:
model0 <- lm(Sepal.Length ~ ., data = iris[, -5])
model <- pmml(model0)

# The resulting PMML:
# <PMML version="4.3" ... xmlns="http://www.dmg.org/PMML-4_3">
# <Header ... description="Linear Regression Model"/>
# <DataDictionary numberOfFields="4">
# .
# .
# </DataDictionary>
# <RegressionModel modelName="Linear_Regression_Model"
#     functionName="regression"
#     algorithmName="least squares">
#  <MiningSchema>
#  .
#  .
#  </MiningSchema>
# .
# .
#  <RegressionTable intercept="1.85599749291755">
#   <NumericPredictor name="Sepal.Width" exponent="1"
#    coefficient="0.650837159313218"/>
#   <NumericPredictor name="Petal.Length" exponent="1"
#    coefficient="0.709131959136729"/>
#   <NumericPredictor name="Petal.Width" exponent="1"
#    coefficient="-0.556482660167024"/>
#  </RegressionTable>
# </RegressionModel>
# </PMML>

# Add arbitrary attributes to the 1st 'NumericPredictor' element. The
# attributes are for demostration only (they are not allowed under
# the PMML schema). The command assumes the default namespace.
add_attributes(model, "/p:PMML/descendant::p:NumericPredictor[1]",
  attributes = c(a = 1, b = "b")
)

#> <?xml version="1.0"?>
#> <PMML xmlns="http://www.dmg.org/PMML-4_3" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" version="4.3" xsi:schemaLocation="http://www.dmg.org/PMML-4_3 http://www.dmg.org/pmml/v4-3/pmml-4-3.xsd">
#>   <Header copyright="Copyright (c) 2019 dmitriy" description="Linear Regression Model">
#>     <Extension name="user" value="dmitriy" extender="SoftwareAG PMML Generator"/>
#>     <Application name="SoftwareAG PMML Generator" version="1.4"/>
#>     <Timestamp>2019-05-29 16:57:27</Timestamp>
#>   </Header>
#>   <DataDictionary numberOfFields="4">
#>     <DataField name="Sepal.Length" optype="continuous" dataType="double"/>
#>     <DataField name="Sepal.Width" optype="continuous" dataType="double"/>
#>     <DataField name="Petal.Length" optype="continuous" dataType="double"/>
#>     <DataField name="Petal.Width" optype="continuous" dataType="double"/>
#>   </DataDictionary>
#>   <RegressionModel modelName="lm_Model" functionName="regression" algorithmName="least squares">
#>     <MiningSchema>
#>       <MiningField name="Sepal.Length" usageType="predicted" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Sepal.Width" usageType="active" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Petal.Length" usageType="active" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Petal.Width" usageType="active" invalidValueTreatment="returnInvalid"/>
#>     </MiningSchema>
#>     <Output>
#>       <OutputField name="Predicted_Sepal.Length" optype="continuous" dataType="double" feature="predictedValue"/>
#>     </Output>
#>     <RegressionTable intercept="1.85599749291754">
#>       <NumericPredictor name="Sepal.Width" exponent="1" coefficient="0.650837159313219" a="1" b="b"/>
#>       <NumericPredictor name="Petal.Length" exponent="1" coefficient="0.709131959136729"/>
#>       <NumericPredictor name="Petal.Width" exponent="1" coefficient="-0.556482660167025"/>
#>     </RegressionTable>
#>   </RegressionModel>
#> </PMML>
#>  

# Add attributes to the NumericPredictor element which has
# 'Petal.Length' as the 'name' attribute:
add_attributes(model,
  "/p:PMML/descendant::p:NumericPredictor[@name='Petal.Length']",
  attributes = c(a = 1, b = "b")
)

#> <?xml version="1.0"?>
#> <PMML xmlns="http://www.dmg.org/PMML-4_3" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" version="4.3" xsi:schemaLocation="http://www.dmg.org/PMML-4_3 http://www.dmg.org/pmml/v4-3/pmml-4-3.xsd">
#>   <Header copyright="Copyright (c) 2019 dmitriy" description="Linear Regression Model">
#>     <Extension name="user" value="dmitriy" extender="SoftwareAG PMML Generator"/>
#>     <Application name="SoftwareAG PMML Generator" version="1.4"/>
#>     <Timestamp>2019-05-29 16:57:27</Timestamp>
#>   </Header>
#>   <DataDictionary numberOfFields="4">
#>     <DataField name="Sepal.Length" optype="continuous" dataType="double"/>
#>     <DataField name="Sepal.Width" optype="continuous" dataType="double"/>
#>     <DataField name="Petal.Length" optype="continuous" dataType="double"/>
#>     <DataField name="Petal.Width" optype="continuous" dataType="double"/>
#>   </DataDictionary>
#>   <RegressionModel modelName="lm_Model" functionName="regression" algorithmName="least squares">
#>     <MiningSchema>
#>       <MiningField name="Sepal.Length" usageType="predicted" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Sepal.Width" usageType="active" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Petal.Length" usageType="active" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Petal.Width" usageType="active" invalidValueTreatment="returnInvalid"/>
#>     </MiningSchema>
#>     <Output>
#>       <OutputField name="Predicted_Sepal.Length" optype="continuous" dataType="double" feature="predictedValue"/>
#>     </Output>
#>     <RegressionTable intercept="1.85599749291754">
#>       <NumericPredictor name="Sepal.Width" exponent="1" coefficient="0.650837159313219"/>
#>       <NumericPredictor name="Petal.Length" exponent="1" coefficient="0.709131959136729" a="1" b="b"/>
#>       <NumericPredictor name="Petal.Width" exponent="1" coefficient="-0.556482660167025"/>
#>     </RegressionTable>
#>   </RegressionModel>
#> </PMML>
#>  

# 3 NumericElements exist which have '1' as the 'exponent' attribute.
# Add new attributes to the 3rd one:
add_attributes(model,
  "/p:PMML/descendant::p:NumericPredictor[@exponent='1'][3]",
  attributes = c(a = 1, b = "b")
)

#> <?xml version="1.0"?>
#> <PMML xmlns="http://www.dmg.org/PMML-4_3" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" version="4.3" xsi:schemaLocation="http://www.dmg.org/PMML-4_3 http://www.dmg.org/pmml/v4-3/pmml-4-3.xsd">
#>   <Header copyright="Copyright (c) 2019 dmitriy" description="Linear Regression Model">
#>     <Extension name="user" value="dmitriy" extender="SoftwareAG PMML Generator"/>
#>     <Application name="SoftwareAG PMML Generator" version="1.4"/>
#>     <Timestamp>2019-05-29 16:57:27</Timestamp>
#>   </Header>
#>   <DataDictionary numberOfFields="4">
#>     <DataField name="Sepal.Length" optype="continuous" dataType="double"/>
#>     <DataField name="Sepal.Width" optype="continuous" dataType="double"/>
#>     <DataField name="Petal.Length" optype="continuous" dataType="double"/>
#>     <DataField name="Petal.Width" optype="continuous" dataType="double"/>
#>   </DataDictionary>
#>   <RegressionModel modelName="lm_Model" functionName="regression" algorithmName="least squares">
#>     <MiningSchema>
#>       <MiningField name="Sepal.Length" usageType="predicted" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Sepal.Width" usageType="active" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Petal.Length" usageType="active" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Petal.Width" usageType="active" invalidValueTreatment="returnInvalid"/>
#>     </MiningSchema>
#>     <Output>
#>       <OutputField name="Predicted_Sepal.Length" optype="continuous" dataType="double" feature="predictedValue"/>
#>     </Output>
#>     <RegressionTable intercept="1.85599749291754">
#>       <NumericPredictor name="Sepal.Width" exponent="1" coefficient="0.650837159313219"/>
#>       <NumericPredictor name="Petal.Length" exponent="1" coefficient="0.709131959136729"/>
#>       <NumericPredictor name="Petal.Width" exponent="1" coefficient="-0.556482660167025" a="1" b="b"/>
#>     </RegressionTable>
#>   </RegressionModel>
#> </PMML>
#>  

# Add attributes to the 1st element whose 'name' attribute contains
# 'Length':
add_attributes(model,
  "/p:PMML/descendant::p:NumericPredictor[contains(@name,'Length')]",
  attributes = c(a = 1, b = "b")
)

#> <?xml version="1.0"?>
#> <PMML xmlns="http://www.dmg.org/PMML-4_3" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" version="4.3" xsi:schemaLocation="http://www.dmg.org/PMML-4_3 http://www.dmg.org/pmml/v4-3/pmml-4-3.xsd">
#>   <Header copyright="Copyright (c) 2019 dmitriy" description="Linear Regression Model">
#>     <Extension name="user" value="dmitriy" extender="SoftwareAG PMML Generator"/>
#>     <Application name="SoftwareAG PMML Generator" version="1.4"/>
#>     <Timestamp>2019-05-29 16:57:27</Timestamp>
#>   </Header>
#>   <DataDictionary numberOfFields="4">
#>     <DataField name="Sepal.Length" optype="continuous" dataType="double"/>
#>     <DataField name="Sepal.Width" optype="continuous" dataType="double"/>
#>     <DataField name="Petal.Length" optype="continuous" dataType="double"/>
#>     <DataField name="Petal.Width" optype="continuous" dataType="double"/>
#>   </DataDictionary>
#>   <RegressionModel modelName="lm_Model" functionName="regression" algorithmName="least squares">
#>     <MiningSchema>
#>       <MiningField name="Sepal.Length" usageType="predicted" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Sepal.Width" usageType="active" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Petal.Length" usageType="active" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Petal.Width" usageType="active" invalidValueTreatment="returnInvalid"/>
#>     </MiningSchema>
#>     <Output>
#>       <OutputField name="Predicted_Sepal.Length" optype="continuous" dataType="double" feature="predictedValue"/>
#>     </Output>
#>     <RegressionTable intercept="1.85599749291754">
#>       <NumericPredictor name="Sepal.Width" exponent="1" coefficient="0.650837159313219"/>
#>       <NumericPredictor name="Petal.Length" exponent="1" coefficient="0.709131959136729" a="1" b="b"/>
#>       <NumericPredictor name="Petal.Width" exponent="1" coefficient="-0.556482660167025"/>
#>     </RegressionTable>
#>   </RegressionModel>
#> </PMML>
#>