pmml.arules {pmml}	R Documentation

Generate PMML for arules objects

Description

Generate the PMML (Predictive Model Markup Language) representation of a rules or an itemset object from package arules. The PMML can then be imported into other systems that accept PMML.

Usage

## S3 method for class 'rules':
pmml(model, model.name="arules_Model", 
    app.name="Rattle/PMML",
    description="arules association rules model", copyright=NULL, ...)
## S3 method for class 'itemsets':
pmml(model, model.name="arules_Model", 
    app.name="Rattle/PMML",
    description="arules frequent itemsets model", copyright=NULL, ...)

Arguments

model 	 an rules or itemsets object.

model.name 	 a name to give to the model in the PMML.

app.name 	 the name of the application that generated the PMML.

description 	 a descriptive text for the header of the PMML.

copyright 	 the copyright notice for the model.

... 	 further arguments passed to or from other methods.

Details

The generated PMML can be imported into any PMML consuming application.

Author(s)

Michael Hahsler (michael@hahsler.net)

References

Package arules home page: http://r-forge.r-project.org/projects/arules

Package home page: http://rattle.togaware.com

PMML home page: http://www.dmg.org

See Also

pmml.

[Package pmml version 1.1.9 Index]