pmml.lm {pmml}	R Documentation

Generate PMML for an lm object

Description

Generate the PMML (Predictive Model Markup Language) representation of an lm object. The PMML can then be imported into other systems that accept PMML.

Usage

## S3 method for class 'lm':
pmml(model, model.name="Regression_model", app.name="Rattle/PMML",
     description="Regression Model", copyright=NULL, ...)

Arguments

model 	 an lm object.

model.name 	 a name to give to the model in the PMML.

app.name 	 the name of the application that generated the PMML.

description 	 a descriptive text for the header of the PMML.

copyright 	 the copyright notice for the model.

... 	 further arguments passed to or from other methods.

Details

The generated PMML can be imported into any PMML consuming application, such as Zementis' ADAPA.

Currently, the resultant PMML document will not encode interaction terms.

Only numeric regression is supported currently. Generalised linear models (logistic regression) are not yet supported.

Author(s)

rguha@indiana.edu

References

Package home page: http://rattle.togaware.com

PMML home page: http://www.dmg.org

Zementis' useful PMML convert: http://www.zementis.com/pmml_converters.htm

See Also

pmml.

[Package pmml version 1.1.9 Index]