Generate the PMML representation for a nnet object from package nnet.

Generate the PMML representation for a nnet object from package nnet.

# S3 method for nnet
pmml(
  model,
  model_name = "NeuralNet_model",
  app_name = "SoftwareAG PMML Generator",
  description = "Neural Network Model",
  copyright = NULL,
  model_version = NULL,
  transforms = NULL,
  missing_value_replacement = NULL,
  ...
)

Arguments

model
A nnet object.

model_name
A name to be given to the PMML model.

app_name
The name of the application that generated the PMML.

description
A descriptive text for the Header element of the PMML.

copyright
The copyright notice for the model.

model_version
A string specifying the model version.

transforms
Data transformations.

missing_value_replacement
Value to be used as the 'missingValueReplacement' attribute for all MiningFields.

...
Further arguments passed to or from other methods.

Value

PMML representation of the nnet object.

Details

This function supports both regression and classification neural network models. The model is represented in the PMML NeuralNetwork format.

References

nnet: Feed-forward Neural Networks and Multinomial Log-Linear Models (on CRAN)

Author

Tridivesh Jena

Examples

library(nnet)
fit <- nnet(Species ~ ., data = iris, size = 4)
#> # weights:  35
#> initial  value 168.526082 
#> iter  10 value 66.525796
#> iter  20 value 41.928309
#> iter  30 value 7.472947
#> iter  40 value 6.153270
#> iter  50 value 5.461043
#> iter  60 value 3.645428
#> iter  70 value 3.528160
#> iter  80 value 3.268686
#> iter  90 value 3.182114
#> iter 100 value 3.065142
#> final  value 3.065142 
#> stopped after 100 iterations
fit_pmml <- pmml(fit)

rm(fit)