Generate the PMML representation for a randomForest object from the package randomForest.

Generate the PMML representation for a randomForest object from the package randomForest.

# S3 method for randomForest
pmml(
  model,
  model_name = "randomForest_Model",
  app_name = "SoftwareAG PMML Generator",
  description = "Random Forest Tree Model",
  copyright = NULL,
  transforms = NULL,
  missing_value_replacement = NULL,
  parent_invalid_value_treatment = "returnInvalid",
  child_invalid_value_treatment = "asIs",
  ...
)

Arguments

model
A randomForest object.

model_name
A name to be given to the PMML model.

app_name
The name of the application that generated the PMML.

description
A descriptive text for the Header element of the PMML.

copyright
The copyright notice for the model.

transforms
Data transformations.

missing_value_replacement
Value to be used as the 'missingValueReplacement' attribute for all MiningFields.

parent_invalid_value_treatment
Invalid value treatment at the top MiningField level.

child_invalid_value_treatment
Invalid value treatment at the model segment MiningField level.

...
Further arguments passed to or from other methods.

Value

PMML representation of the randomForest object.

Details

This function outputs a Random Forest in PMML format.

References

randomForest: Breiman and Cutler's random forests for classification and regression

Author

Tridivesh Jena

Examples

# Build a randomForest model
library(randomForest)

#> randomForest 4.6-14

#> Type rfNews() to see new features/changes/bug fixes.

#> 
#> Attaching package: 'randomForest'

#> The following object is masked from 'package:rattle':
#> 
#>     importance

iris_rf <- randomForest(Species ~ ., data = iris, ntree = 20)

# Convert to pmml
iris_rf_pmml <- pmml(iris_rf)

#> Now converting tree  1  to PMML

#> Now converting tree  2  to PMML

#> Now converting tree  3  to PMML

#> Now converting tree  4  to PMML

#> Now converting tree  5  to PMML

#> Now converting tree  6  to PMML

#> Now converting tree  7  to PMML

#> Now converting tree  8  to PMML

#> Now converting tree  9  to PMML

#> Now converting tree  10  to PMML

#> Now converting tree  11  to PMML

#> Now converting tree  12  to PMML

#> Now converting tree  13  to PMML

#> Now converting tree  14  to PMML

#> Now converting tree  15  to PMML

#> Now converting tree  16  to PMML

#> Now converting tree  17  to PMML

#> Now converting tree  18  to PMML

#> Now converting tree  19  to PMML

#> Now converting tree  20  to PMML

rm(iris_rf)