pmml.rsf {pmml}	R Documentation

Generate PMML for a Random Survival Forest (rsf) object

Description

Generate the Predictive Model Markup Language (PMML) representation of a randomSurvivalForest forest object. In particular, this function gives the user the ability to save the geometry of a forest as a PMML XML document.

Usage

## S3 method for class 'rsf':
pmml(model, model.name="rsfForest_Model", app.name="Rattle/PMML",
     description="Random Survival Forest Tree Model", copyright, ...)

Arguments

model 	 the forest object contained in an object of class randomSurvivalForest, as that contained in the object returned by the function rsf with the parameter “forest=TRUE”.

model.name 	 a name to give to the model in the PMML.

app.name 	 the name of the application that generated the PMML.

description 	 a descriptive text for the header of the PMML.

copyright 	 the copyright notice for the model.

... 	 further arguments passed to or from other methods.

Details

The Predictive Model Markup Language is an XML based language which provides a way for applications to define statistical and data mining models and to share models between PMML compliant applications. More information about PMML and the Data Mining Group can be found at http://www.dmg.org.

Use of PMML and pmml.rsf requires the XML package. Be aware that XML is a very verbose data format. Reasonably sized trees and data sets can lead to extremely large text files. XML, while achieving interoperability, is not an efficient data storage mechanism in this case.

This function is used to export the geometry of the forest to other PMML compliant applications, including graphics packages that are capable of printing binary trees. In addition, the user may wish to save the geometry of the forest for later retrieval and prediction on new data sets using pmml.rsf together with pmml_to_rsf.

Value

An object of class XMLNode as that defined by the XML package. This represents the top level, or root node, of the XML document and is of type PMML. It can be written to file with saveXML.

Note

One cautionary note is in order. The PMML representation of the randomSurvivalForest forest object is incomplete, in that the object needs to be massaged in order for prediction to be possible. This will be clear in the examples. This deficiency will be addressed in future releases of this package. However, it was felt that the current functionality was important enough and mature enough to warrant release in this version of the product.

Author(s)

Hemant Ishwaran hemant.ishwaran@gmail.com and Udaya B. Kogalur ubk2101@columbia.edu with incorporation into the pmml package by Graham.Williams@togaware.com

References

H. Ishwaran and Udaya B. Kogalur (2006). Random Survival Forests. Cleveland Clinic Technical Report.

PMML home page: http://www.dmg.org

See Also

pmml.

Examples

library(randomSurvivalForest)
data(veteran, package = "randomSurvivalForest")
veteran.out <- rsf(Survrsf(time, status)~.,
        data = veteran,
        ntree = 5,
        forest = TRUE)
veteran.forest <- veteran.out$forest
pmml(veteran.forest)

[Package pmml version 1.1.9 Index]