Add attribute values to an existing DataField element in a given PMML file

Add attribute values to an existing DataField element in a given PMML file

add_data_field_attributes(xml_model = NULL, attributes = NULL,
  field = NULL, namespace = "4_3", ...)

Arguments

xml_model
The PMML model in a XML node format. If the model is a text file, it should be converted to an XML node, for example, using the file_to_xml_node function.

attributes
The attributes to be added to the data fields. The user should make sure that the attributes being added are allowed in the PMML schema.

field
The field to which the attributes are to be added. This is used when the attributes are a vector of name-value pairs, intended for this one field.

namespace
The namespace of the PMML model. This is frequently also the PMML version of the model.

…
Further arguments passed to or from other methods.

Value

An object of class XMLNode as that defined by the XML package. This represents the top level, or root node, of the XML document and is of type PMML. It can be written to file with saveXML.

Details

The PMML format allows a DataField element to have various attributes, which, although useful, may not always be present in a PMML model. This function allows one to take an existing PMML file and add these attributes to the DataFields.

The attribute information can be provided as a dataframe or a vector. Each row of the data frame corresponds to an attribute name and each column corresponding to a variable name. This way one can add as many attributes to as many variables as one wants in one step. A more convinient method to add multiple attributes to one field might be to give the attribute name and values as a vector. This function may be used multiple times to add new attribute values step-by-step. However this function overwrites any pre-existing attribute values, so it must be used with care. This behavior is by design as this feature is meant to help an user add new defined attribute values at different times. For example, one may use this to modify the display name of a field at different times.

Examples

# Make a sample model:
model0 <- lm(Sepal.Length ~ ., data = iris[, -5])
model <- pmml(model0)

# The resulting model has mining fields with no information besides
# fieldName, dataType and optype. This object is already an xml
# node (not an external text file), so there is no need to convert
# it to an xml node object.

# Create data frame with attribute information:

attributes <- data.frame(c("FlowerWidth", 1), c("FlowerLength", 0),
  stringAsFactors = FALSE
)
rownames(attributes) <- c("displayName", "isCyclic")
colnames(attributes) <- c("Sepal.Width", "Petal.Length")

# Although not needed in this first try, necessary to easily add
# new values later. Removes values as factors so that new values
# added later are not evaluated as factor values and thus rejected
# as invalid.
attributes[] <- lapply(attributes, as.character)

add_data_field_attributes(model, attributes, namespace = "4_3")

#> [1] "WARNING: the following field additions will be ignored"
#> NA not in model

#> <?xml version="1.0"?>
#> <PMML xmlns="http://www.dmg.org/PMML-4_3" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" version="4.3" xsi:schemaLocation="http://www.dmg.org/PMML-4_3 http://www.dmg.org/pmml/v4-3/pmml-4-3.xsd">
#>   <Header copyright="Copyright (c) 2019 dmitriy" description="Linear Regression Model">
#>     <Extension name="user" value="dmitriy" extender="SoftwareAG PMML Generator"/>
#>     <Application name="SoftwareAG PMML Generator" version="1.4"/>
#>     <Timestamp>2019-06-04 15:20:29</Timestamp>
#>   </Header>
#>   <DataDictionary numberOfFields="4">
#>     <DataField name="Sepal.Length" optype="continuous" dataType="double"/>
#>     <DataField name="Sepal.Width" optype="continuous" dataType="double" displayName="FlowerWidth" isCyclic="1"/>
#>     <DataField name="Petal.Length" optype="continuous" dataType="double" displayName="FlowerLength" isCyclic="0"/>
#>     <DataField name="Petal.Width" optype="continuous" dataType="double"/>
#>   </DataDictionary>
#>   <RegressionModel modelName="lm_Model" functionName="regression" algorithmName="least squares">
#>     <MiningSchema>
#>       <MiningField name="Sepal.Length" usageType="predicted" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Sepal.Width" usageType="active" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Petal.Length" usageType="active" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Petal.Width" usageType="active" invalidValueTreatment="returnInvalid"/>
#>     </MiningSchema>
#>     <Output>
#>       <OutputField name="Predicted_Sepal.Length" optype="continuous" dataType="double" feature="predictedValue"/>
#>     </Output>
#>     <RegressionTable intercept="1.85599749291754">
#>       <NumericPredictor name="Sepal.Width" exponent="1" coefficient="0.650837159313219"/>
#>       <NumericPredictor name="Petal.Length" exponent="1" coefficient="0.709131959136729"/>
#>       <NumericPredictor name="Petal.Width" exponent="1" coefficient="-0.556482660167025"/>
#>     </RegressionTable>
#>   </RegressionModel>
#> </PMML>
#>  

# Alternative method to add attributes to a single field,
# "Sepal.Width":
add_data_field_attributes(
  model, c(displayName = "FlowerWidth", isCyclic = 1),
  "Sepal.Width"
)

#> <?xml version="1.0"?>
#> <PMML xmlns="http://www.dmg.org/PMML-4_3" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" version="4.3" xsi:schemaLocation="http://www.dmg.org/PMML-4_3 http://www.dmg.org/pmml/v4-3/pmml-4-3.xsd">
#>   <Header copyright="Copyright (c) 2019 dmitriy" description="Linear Regression Model">
#>     <Extension name="user" value="dmitriy" extender="SoftwareAG PMML Generator"/>
#>     <Application name="SoftwareAG PMML Generator" version="1.4"/>
#>     <Timestamp>2019-06-04 15:20:29</Timestamp>
#>   </Header>
#>   <DataDictionary numberOfFields="4">
#>     <DataField name="Sepal.Length" optype="continuous" dataType="double"/>
#>     <DataField name="Sepal.Width" optype="continuous" dataType="double" displayName="FlowerWidth" isCyclic="1"/>
#>     <DataField name="Petal.Length" optype="continuous" dataType="double"/>
#>     <DataField name="Petal.Width" optype="continuous" dataType="double"/>
#>   </DataDictionary>
#>   <RegressionModel modelName="lm_Model" functionName="regression" algorithmName="least squares">
#>     <MiningSchema>
#>       <MiningField name="Sepal.Length" usageType="predicted" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Sepal.Width" usageType="active" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Petal.Length" usageType="active" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Petal.Width" usageType="active" invalidValueTreatment="returnInvalid"/>
#>     </MiningSchema>
#>     <Output>
#>       <OutputField name="Predicted_Sepal.Length" optype="continuous" dataType="double" feature="predictedValue"/>
#>     </Output>
#>     <RegressionTable intercept="1.85599749291754">
#>       <NumericPredictor name="Sepal.Width" exponent="1" coefficient="0.650837159313219"/>
#>       <NumericPredictor name="Petal.Length" exponent="1" coefficient="0.709131959136729"/>
#>       <NumericPredictor name="Petal.Width" exponent="1" coefficient="-0.556482660167025"/>
#>     </RegressionTable>
#>   </RegressionModel>
#> </PMML>
#>  

mi <- make_intervals(
  list("openClosed", "closedClosed", "closedOpen"),
  list(NULL, 1, 2), list(1, 2, NULL)
)
mv <- make_values(
  list("A", "B", "C"), list(NULL, NULL, NULL),
  list("valid", NULL, "invalid")
)
add_data_field_children(model, field = "Sepal.Length", interval = mi, values = mv)

#> <?xml version="1.0"?>
#> <PMML xmlns="http://www.dmg.org/PMML-4_3" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" version="4.3" xsi:schemaLocation="http://www.dmg.org/PMML-4_3 http://www.dmg.org/pmml/v4-3/pmml-4-3.xsd">
#>   <Header copyright="Copyright (c) 2019 dmitriy" description="Linear Regression Model">
#>     <Extension name="user" value="dmitriy" extender="SoftwareAG PMML Generator"/>
#>     <Application name="SoftwareAG PMML Generator" version="1.4"/>
#>     <Timestamp>2019-06-04 15:20:29</Timestamp>
#>   </Header>
#>   <DataDictionary numberOfFields="4">
#>     <DataField name="Sepal.Length" optype="continuous" dataType="double">
#>       <Interval closure="openClosed" rightMargin="1"/>
#>       <Interval closure="closedClosed" leftMargin="1" rightMargin="2"/>
#>       <Interval closure="closedOpen" leftMargin="2"/>
#>       <Value value="A" property="valid"/>
#>       <Value value="B"/>
#>       <Value value="C" property="invalid"/>
#>     </DataField>
#>     <DataField name="Sepal.Width" optype="continuous" dataType="double"/>
#>     <DataField name="Petal.Length" optype="continuous" dataType="double"/>
#>     <DataField name="Petal.Width" optype="continuous" dataType="double"/>
#>   </DataDictionary>
#>   <RegressionModel modelName="lm_Model" functionName="regression" algorithmName="least squares">
#>     <MiningSchema>
#>       <MiningField name="Sepal.Length" usageType="predicted" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Sepal.Width" usageType="active" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Petal.Length" usageType="active" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Petal.Width" usageType="active" invalidValueTreatment="returnInvalid"/>
#>     </MiningSchema>
#>     <Output>
#>       <OutputField name="Predicted_Sepal.Length" optype="continuous" dataType="double" feature="predictedValue"/>
#>     </Output>
#>     <RegressionTable intercept="1.85599749291754">
#>       <NumericPredictor name="Sepal.Width" exponent="1" coefficient="0.650837159313219"/>
#>       <NumericPredictor name="Petal.Length" exponent="1" coefficient="0.709131959136729"/>
#>       <NumericPredictor name="Petal.Width" exponent="1" coefficient="-0.556482660167025"/>
#>     </RegressionTable>
#>   </RegressionModel>
#> </PMML>
#>