Add 'Interval' and 'Value' child elements to a given DataField element in a given PMML file.

Add 'Interval' and 'Value' child elements to a given DataField element in a given PMML file.

add_data_field_children(xml_model = NULL, field = NULL,
  intervals = NULL, values = NULL, namespace = "4_3", ...)

Arguments

xml_model
The PMML model in a XML node format. If the model is a text file, it should be converted to an XML node, for example, using the file_to_xml_node function.

field
The field to which the attributes are to be added. This is used when the attributes are a vector of name-value pairs, intended for this one field.

intervals
The 'Interval' elements given as a list

values
The 'Value' elements given as a list.

namespace
The namespace of the PMML model. This is frequently also the PMML version of the model.

…
Further arguments passed to or from other methods.

Value

An object of class XMLNode as that defined by the XML package. This represents the top level, or root node, of the XML document and is of type PMML. It can be written to file with saveXML.

Details

The PMML format allows a DataField element to have 'Interval' and 'Value' child elements which although useful, may not always be present in a PMML model. This function allows one to take an existing PMML file and add these elements to the DataFields.

The 'Interval' elements or the 'Value' elements can be typed in, but more conviniently created by using the helper functions 'make_intervals' and 'MakeValues'. This function can then add these extra information to the PMML.

Examples

# Make a sample model:
model0 <- lm(Sepal.Length ~ ., data = iris[, -5])
model <- pmml(model0)

# The resulting model has data fields but with no 'Interval' or Value'
# elements. This object is already an xml node (not an external text
# file), so there is no need to convert it to an xml node object.

# Add an 'Interval' element node by typing it in
add_data_field_children(model,
  field = "Sepal.Length",
  intervals = list(newXMLNode("Interval",
    attrs = c(closure = "openClosed", rightMargin = 3)
  ))
)

#> <?xml version="1.0"?>
#> <PMML xmlns="http://www.dmg.org/PMML-4_3" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" version="4.3" xsi:schemaLocation="http://www.dmg.org/PMML-4_3 http://www.dmg.org/pmml/v4-3/pmml-4-3.xsd">
#>   <Header copyright="Copyright (c) 2019 dmitriy" description="Linear Regression Model">
#>     <Extension name="user" value="dmitriy" extender="SoftwareAG PMML Generator"/>
#>     <Application name="SoftwareAG PMML Generator" version="1.4"/>
#>     <Timestamp>2019-05-29 16:21:42</Timestamp>
#>   </Header>
#>   <DataDictionary numberOfFields="4">
#>     <DataField name="Sepal.Length" optype="continuous" dataType="double">
#>       <Interval closure="openClosed" rightMargin="3"/>
#>     </DataField>
#>     <DataField name="Sepal.Width" optype="continuous" dataType="double"/>
#>     <DataField name="Petal.Length" optype="continuous" dataType="double"/>
#>     <DataField name="Petal.Width" optype="continuous" dataType="double"/>
#>   </DataDictionary>
#>   <RegressionModel modelName="lm_Model" functionName="regression" algorithmName="least squares">
#>     <MiningSchema>
#>       <MiningField name="Sepal.Length" usageType="predicted" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Sepal.Width" usageType="active" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Petal.Length" usageType="active" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Petal.Width" usageType="active" invalidValueTreatment="returnInvalid"/>
#>     </MiningSchema>
#>     <Output>
#>       <OutputField name="Predicted_Sepal.Length" optype="continuous" dataType="double" feature="predictedValue"/>
#>     </Output>
#>     <RegressionTable intercept="1.85599749291754">
#>       <NumericPredictor name="Sepal.Width" exponent="1" coefficient="0.650837159313219"/>
#>       <NumericPredictor name="Petal.Length" exponent="1" coefficient="0.709131959136729"/>
#>       <NumericPredictor name="Petal.Width" exponent="1" coefficient="-0.556482660167025"/>
#>     </RegressionTable>
#>   </RegressionModel>
#> </PMML>
#>  

# Use helper functions to create list of 'Interval' and 'Value'
# elements. We define the 3 Intervals as ,1]  (1,2)  and [2,
mi <- make_intervals(
  list("openClosed", "openOpen", "closedOpen"),
  list(NULL, 1, 2), list(1, 2, NULL)
)

# Define 3 values, none with a 'displayValue' attribute and 1 value
# defined as 'invalid'. The 2nd one is 'valid' by default.
mv <- make_values(
  list(1.1, 2.2, 3.3), list(NULL, NULL, NULL),
  list("valid", NULL, "invalid")
)

# As an example, apply these to the Sepal.Length field:
add_data_field_children(model, field = "Sepal.Length", intervals = mi, values = mv)

#> <?xml version="1.0"?>
#> <PMML xmlns="http://www.dmg.org/PMML-4_3" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" version="4.3" xsi:schemaLocation="http://www.dmg.org/PMML-4_3 http://www.dmg.org/pmml/v4-3/pmml-4-3.xsd">
#>   <Header copyright="Copyright (c) 2019 dmitriy" description="Linear Regression Model">
#>     <Extension name="user" value="dmitriy" extender="SoftwareAG PMML Generator"/>
#>     <Application name="SoftwareAG PMML Generator" version="1.4"/>
#>     <Timestamp>2019-05-29 16:21:42</Timestamp>
#>   </Header>
#>   <DataDictionary numberOfFields="4">
#>     <DataField name="Sepal.Length" optype="continuous" dataType="double">
#>       <Interval closure="openClosed" rightMargin="1"/>
#>       <Interval closure="openOpen" leftMargin="1" rightMargin="2"/>
#>       <Interval closure="closedOpen" leftMargin="2"/>
#>       <Value value="1.1" property="valid"/>
#>       <Value value="2.2"/>
#>       <Value value="3.3" property="invalid"/>
#>     </DataField>
#>     <DataField name="Sepal.Width" optype="continuous" dataType="double"/>
#>     <DataField name="Petal.Length" optype="continuous" dataType="double"/>
#>     <DataField name="Petal.Width" optype="continuous" dataType="double"/>
#>   </DataDictionary>
#>   <RegressionModel modelName="lm_Model" functionName="regression" algorithmName="least squares">
#>     <MiningSchema>
#>       <MiningField name="Sepal.Length" usageType="predicted" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Sepal.Width" usageType="active" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Petal.Length" usageType="active" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Petal.Width" usageType="active" invalidValueTreatment="returnInvalid"/>
#>     </MiningSchema>
#>     <Output>
#>       <OutputField name="Predicted_Sepal.Length" optype="continuous" dataType="double" feature="predictedValue"/>
#>     </Output>
#>     <RegressionTable intercept="1.85599749291754">
#>       <NumericPredictor name="Sepal.Width" exponent="1" coefficient="0.650837159313219"/>
#>       <NumericPredictor name="Petal.Length" exponent="1" coefficient="0.709131959136729"/>
#>       <NumericPredictor name="Petal.Width" exponent="1" coefficient="-0.556482660167025"/>
#>     </RegressionTable>
#>   </RegressionModel>
#> </PMML>
#>  

# Only defined 'Interval's:
add_data_field_children(model, field = "Sepal.Length", intervals = mi)

#> <?xml version="1.0"?>
#> <PMML xmlns="http://www.dmg.org/PMML-4_3" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" version="4.3" xsi:schemaLocation="http://www.dmg.org/PMML-4_3 http://www.dmg.org/pmml/v4-3/pmml-4-3.xsd">
#>   <Header copyright="Copyright (c) 2019 dmitriy" description="Linear Regression Model">
#>     <Extension name="user" value="dmitriy" extender="SoftwareAG PMML Generator"/>
#>     <Application name="SoftwareAG PMML Generator" version="1.4"/>
#>     <Timestamp>2019-05-29 16:21:42</Timestamp>
#>   </Header>
#>   <DataDictionary numberOfFields="4">
#>     <DataField name="Sepal.Length" optype="continuous" dataType="double">
#>       <default:Interval xmlns:default="http://www.dmg.org/PMML-4_3" closure="openClosed" rightMargin="1"/>
#>       <default:Interval xmlns:default="http://www.dmg.org/PMML-4_3" closure="openOpen" leftMargin="1" rightMargin="2"/>
#>       <default:Interval xmlns:default="http://www.dmg.org/PMML-4_3" closure="closedOpen" leftMargin="2"/>
#>     </DataField>
#>     <DataField name="Sepal.Width" optype="continuous" dataType="double"/>
#>     <DataField name="Petal.Length" optype="continuous" dataType="double"/>
#>     <DataField name="Petal.Width" optype="continuous" dataType="double"/>
#>   </DataDictionary>
#>   <RegressionModel modelName="lm_Model" functionName="regression" algorithmName="least squares">
#>     <MiningSchema>
#>       <MiningField name="Sepal.Length" usageType="predicted" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Sepal.Width" usageType="active" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Petal.Length" usageType="active" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Petal.Width" usageType="active" invalidValueTreatment="returnInvalid"/>
#>     </MiningSchema>
#>     <Output>
#>       <OutputField name="Predicted_Sepal.Length" optype="continuous" dataType="double" feature="predictedValue"/>
#>     </Output>
#>     <RegressionTable intercept="1.85599749291754">
#>       <NumericPredictor name="Sepal.Width" exponent="1" coefficient="0.650837159313219"/>
#>       <NumericPredictor name="Petal.Length" exponent="1" coefficient="0.709131959136729"/>
#>       <NumericPredictor name="Petal.Width" exponent="1" coefficient="-0.556482660167025"/>
#>     </RegressionTable>
#>   </RegressionModel>
#> </PMML>
#>