Add attribute values to an existing MiningField element in a given PMML file.

Add attribute values to an existing MiningField element in a given PMML file.

add_mining_field_attributes(xml_model = NULL, attributes = NULL,
  namespace = "4_3", ...)

Arguments

xml_model
The PMML model in a XML node format. If the model is a text file, it should be converted to an XML node, for example, using the file_to_xml_node function.

attributes
The attributes to be added to the mining fields. The user should make sure that the attributes being added are allowed in the PMML schema.

namespace
The namespace of the PMML model. This is frequently also the PMML version of the model.

…
Further arguments passed to or from other methods.

Value

An object of class XMLNode as that defined by the XML package. This represents the top level, or root node, of the XML document and is of type PMML. It can be written to file with saveXML.

Details

The PMML format allows a MiningField element to have attributes 'usageType', 'missingValueReplacement' and 'invalidValueTreatment' which although useful, may not always be present in a PMML model. This function allows one to take an existing PMML file and add these attributes to the MiningFields.

The attribute information should be provided as a dataframe; each row corresponding to an attribute name and each column corresponding to a variable name. This way one can add as many attributes to as many variables as one wants in one step. On the other extreme, a one-by-one data frame may be used to add one new attribute to one variable. This function may be used multiple times to add new attribute values step-by-step. This function overwrites any pre-existing attribute values, so it must be used with care. However, this is by design as this feature is meant to help an user defined new attribute values at different times. For example, one may use this to impute missing values in a model at different times.

Examples

# Make a sample model
model0 <- lm(Sepal.Length ~ ., data = iris[, -5])
model <- pmml(model0)

# The resulting model has mining fields with no information
# besides fieldName, dataType and optype. This object is
# already an xml node (not an external text file), so there
# is no need to convert it to an xml node object.

# Create data frame with attribute information:
attributes <- data.frame(
  c("active", 1.1, "asIs"),
  c("active", 2.2, "asIs"),
  c("active", NA, "asMissing")
)
rownames(attributes) <- c(
  "usageType", "missingValueReplacement",
  "invalidValueTreatment"
)
colnames(attributes) <- c(
  "Sepal.Width", "Petal.Length",
  "Petal.Width"
)

# Although not needed in this first try, necessary to easily
# add new values later:
for (k in 1:ncol(attributes)) {
  attributes[[k]] <- as.character(attributes[[k]])
}

add_mining_field_attributes(model, attributes, namespace = "4_3")

#> <?xml version="1.0"?>
#> <PMML xmlns="http://www.dmg.org/PMML-4_3" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" version="4.3" xsi:schemaLocation="http://www.dmg.org/PMML-4_3 http://www.dmg.org/pmml/v4-3/pmml-4-3.xsd">
#>   <Header copyright="Copyright (c) 2019 dmitriy" description="Linear Regression Model">
#>     <Extension name="user" value="dmitriy" extender="SoftwareAG PMML Generator"/>
#>     <Application name="SoftwareAG PMML Generator" version="1.4"/>
#>     <Timestamp>2019-05-30 15:05:20</Timestamp>
#>   </Header>
#>   <DataDictionary numberOfFields="4">
#>     <DataField name="Sepal.Length" optype="continuous" dataType="double"/>
#>     <DataField name="Sepal.Width" optype="continuous" dataType="double"/>
#>     <DataField name="Petal.Length" optype="continuous" dataType="double"/>
#>     <DataField name="Petal.Width" optype="continuous" dataType="double"/>
#>   </DataDictionary>
#>   <RegressionModel modelName="lm_Model" functionName="regression" algorithmName="least squares">
#>     <MiningSchema>
#>       <MiningField name="Sepal.Length" usageType="predicted" invalidValueTreatment="returnInvalid"/>
#>       <MiningField name="Sepal.Width" usageType="active" invalidValueTreatment="asIs" missingValueReplacement="1.1"/>
#>       <MiningField name="Petal.Length" usageType="active" invalidValueTreatment="asIs" missingValueReplacement="2.2"/>
#>       <MiningField name="Petal.Width" usageType="active" invalidValueTreatment="asMissing"/>
#>     </MiningSchema>
#>     <Output>
#>       <OutputField name="Predicted_Sepal.Length" optype="continuous" dataType="double" feature="predictedValue"/>
#>     </Output>
#>     <RegressionTable intercept="1.85599749291754">
#>       <NumericPredictor name="Sepal.Width" exponent="1" coefficient="0.650837159313219"/>
#>       <NumericPredictor name="Petal.Length" exponent="1" coefficient="0.709131959136729"/>
#>       <NumericPredictor name="Petal.Width" exponent="1" coefficient="-0.556482660167025"/>
#>     </RegressionTable>
#>   </RegressionModel>
#> </PMML>
#>