Convert an R expression to PMML.

Convert an R expression to PMML.

function_to_pmml(expr)

Arguments

expr
An R expression enclosed in quotes.

Value

PMML version of the input expression

Details

As long as the expression passed to the function is a valid R expression (e.g., no unbalanced parenthesis), it can contain arbitrary function names not defined in R. Variables in the expression passed to `xform_function` are always assumed to be fields, and not substituted. That is, even if `x` has a value in the R environment, the resulting expression will still use `x`.

An expression such as `foo(x)` is treated as a function `foo` with argument `x`. Consequently, passing in an R vector `c(1,2,3)` to `function_to_pmml()` will produce PMML where `c` is a function and `1,2,3` are the arguments.

An expression starting with '-' or '+' (for example, "-3" or "-(a+b)") will be treated as if there is a 0 before the initial '-' or '+' sign. This makes it possible to represent expressions that start with a sign, since PMML's '-' and '+' functions require two arguments. The resulting PMML node will have a constant 0 as a child.

Examples

# Operator precedence and parenthesis
function_to_pmml("1 + 3/5 - (4 * 2)")

#> <Apply function="-">
#>   <Apply function="+">
#>     <Constant dataType="double">1</Constant>
#>     <Apply function="/">
#>       <Constant dataType="double">3</Constant>
#>       <Constant dataType="double">5</Constant>
#>     </Apply>
#>   </Apply>
#>   <Apply function="*">
#>     <Constant dataType="double">4</Constant>
#>     <Constant dataType="double">2</Constant>
#>   </Apply>
#> </Apply> 

# Nested arbitrary functions
function_to_pmml("foo(bar(x)) - bar(foo(y-z))")

#> <Apply function="-">
#>   <Apply function="foo">
#>     <Apply function="bar">
#>       <FieldRef field="x"/>
#>     </Apply>
#>   </Apply>
#>   <Apply function="bar">
#>     <Apply function="foo">
#>       <Apply function="-">
#>         <FieldRef field="y"/>
#>         <FieldRef field="z"/>
#>       </Apply>
#>     </Apply>
#>   </Apply>
#> </Apply> 

# If-else expression
function_to_pmml("if (x==3) { 3 } else { 0 }")

#> <Apply function="if">
#>   <Apply function="equal">
#>     <FieldRef field="x"/>
#>     <Constant dataType="double">3</Constant>
#>   </Apply>
#>   <Constant dataType="double">3</Constant>
#>   <Constant dataType="double">0</Constant>
#> </Apply> 

# If-else with boolean output
function_to_pmml("if (x==3) { TRUE } else { FALSE }")

#> <Apply function="if">
#>   <Apply function="equal">
#>     <FieldRef field="x"/>
#>     <Constant dataType="double">3</Constant>
#>   </Apply>
#>   <Constant dataType="boolean">TRUE</Constant>
#>   <Constant dataType="boolean">FALSE</Constant>
#> </Apply> 

# Function with string argument types
function_to_pmml("colors('red','green','blue')")

#> <Apply function="colors">
#>   <Constant dataType="string">red</Constant>
#>   <Constant dataType="string">green</Constant>
#>   <Constant dataType="string">blue</Constant>
#> </Apply> 

# Sign in front of expression
function_to_pmml("-(x/y)")

#> <Apply function="-">
#>   <Constant dataType="double">0</Constant>
#>   <Apply function="/">
#>     <FieldRef field="x"/>
#>     <FieldRef field="y"/>
#>   </Apply>
#> </Apply>