pmml {pmml}	R Documentation

Generate PMML for an R object

Description

'pmml' is a generic function implementing S3 methods used to produce the PMML (Predictive Model Markup Language) representation of an R model. The PMML can then be imported into other systems that accept PMML.

Usage

pmml(model, model.name="Rattle_Model", app.name="Rattle/PMML",
     description=NULL, copyright=NULL, ...)

Arguments

model 	 an object to be converted to PMML.

model.name 	 a name to give to the model in the PMML.

app.name 	 the name of the application that generated the PMML.

description 	 a descriptive text for the header of the PMML.

copyright 	 the copyright notice for the model.

... 	 further arguments passed to or from other methods.

Details

The generated PMML can be imported into any PMML consuming application, such as Teradata Warehouse Miner and DB2. Generally, these applications convert the PMML into SQL for execution across a database.

Author(s)

Graham.Williams@togaware.com

References

Package home page: http://rattle.togaware.com

PMML home page: http://www.dmg.org

See Also

pmml.kmeans, pmml.rpart, pmml.rsf.

[Package pmml version 1.1.9 Index]