pmml: Generate PMML for R Models

Export various R models to PMML and generate data transformations in PMML format.

Functions

pmml exports R model objects to PMML format.

xform_wrap creates an object with that can then be used to describe transformations to be exported to PMML.

Data Transformations

The data transformation functions previously available in the separate pmmlTransformations package have been merged into pmml starting with version 2.0.0.

The general methodology is to first wrap the data with xform_wrap, and then perform transformations using the following functions: xform_discretize, xform_function, xform_map, xform_min_max, xform_norm_discrete, xform_z_score. The model, including the transformations, can then be output in PMML format by calling the pmml function. The pmml function in this case has to be given an additional parameter, transforms.

PMML

The Predictive Model Markup Language (PMML) is an XML-based language which provides a way for applications to define machine learning, statistical and data mining models and to share models between PMML compliant applications. More information about the PMML industry standard and the Data Mining Group can be found at <http://www.dmg.org>. The generated PMML can be imported into any PMML consuming application, such as Zementis Predictive Analytics products, which integrate with web services, relational database systems and deploy natively on Hadoop in conjunction with Hive, Spark or Storm, as well as allow predictive analytics to be executed for IBM z Systems mainframe applications and real-time, streaming analytics platforms.

References

PMML home page

PMML transformations

A. Guazzelli, W. Lin, T. Jena (2012), PMML in Action: Unleashing the Power of Open Standards for Data Mining and Predictive Analytics. CreativeSpace (Second Edition) - Available on Amazon.com.

A. Guazzelli, M. Zeller, W. Lin, G. Williams (2009), PMML: An Open Standard for Sharing Models. The R journal, Volume 1/1, 60-65

T. Jena, A. Guazzelli, W. Lin, M. Zeller (2013). The R pmmlTransformations Package. In Proceedings of the 19th ACM SIGKDD Conference on Knowledge Discovery and Data Mining

See also

Author

Maintainer: Dmitriy Bolotov dmitriy.bolotov@softwareag.com

Authors:

Tridivesh Jena tridivesh.jena@softwareag.com

Graham Williams graham.williams@togaware.net

Wen-Ching Lin

Michael Hahsler michael@hahsler.net

Hemant Ishwaran

Udaya B. Kogalur

Rajarshi Guha rguha@indiana.edu

Other contributors:

Software AG [copyright holder]